Changement des types primitifs et ajustement au niveau des tests

+++ td3/src/main/java/com/iut/as/interfaces/IMaths.java
@@ -2,12 +2,12 @@ package com.iut.as.interfaces;
 
 public interface IMaths {
 
-	int addition(int a, int b);
+	Integer addition(Integer a, Integer b);
 
-	int soustraction(int a, int b);
+	Integer soustraction(Integer a, Integer b);
 
 	// int (type primitif) --> Integer (class)
-	double division(Integer a, Integer b);
+	Float division(Integer a, Integer b);
 
-	int multiplication(int a, int b);
+	Integer multiplication(Integer a, Integer b);
 }
